requirements.txt.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu
numpy
sentence-transformers
transformers
google-genai
